Redesigned LCD_Driver interface to be much faster through a new buffer approach that greatly reduces memory overhead and drawing time for screen. Included SD card interface designed by WaveShare, and began implementing own file format for use with buffer and palette interface for simplicity.

+++ lib/lcd/LCD_Input.c
@@ -2,16 +2,16 @@
 
 #include "LCD_Input.h"
 
-uint8_t JoyStick_X() {
+uint16_t JoyStick_X() {
     adc_select_input(0);
-    const uint8_t result = adc_read();
+    const uint16_t result = adc_read();
 
     return result;
 }
 
-uint8_t JoyStick_Y() {
+uint16_t JoyStick_Y() {
     adc_select_input(1);
-    const uint8_t result = adc_read();
+    const uint16_t result = adc_read();
 
     return result;
 }
